requirements.txt
openai-whisper
torch
sounddevice
numpy
openai>=1.0.0
python-dotenv>=1.0.0